% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{add_hyperlinks_to_workbook}
\alias{add_hyperlinks_to_workbook}
\title{Add hyperlinks to gene databases}
\usage{
add_hyperlinks_to_workbook(wb, gene_data)
}
\arguments{
\item{wb}{Workbook object}

\item{gene_data}{Gene data table}
}
\description{
Internal function to add hyperlinks to external resources.
}
\keyword{internal}
