% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGenes.R
\name{applyFilter}
\alias{applyFilter}
\title{Apply a single filter to data}
\usage{
applyFilter(data, filter, verbose = FALSE)
}
\arguments{
\item{data}{Data to filter}

\item{filter}{Filter specification}

\item{verbose}{Whether to print details}
}
\value{
Logical vector
}
\description{
Internal function to apply a structured filter
}
\keyword{internal}
