% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{prioritize_genes}
\alias{prioritize_genes}
\title{Prioritize genes based on multiple criteria}
\usage{
prioritize_genes(gene_data)
}
\arguments{
\item{gene_data}{Gene data table}
}
\value{
Prioritized subset of gene data
}
\description{
Internal function to prioritize genes for the Top Candidates sheet.
}
\keyword{internal}
