% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R, R/zzz_aliases.R
\name{removeRatTable}
\alias{removeRatTable}
\alias{lpr_remove_table}
\title{Remove a Supplementary Table from locusPackRat Project}
\usage{
removeRatTable(table_name, project_dir = ".")

lpr_remove_table(table_name, project_dir = ".")
}
\arguments{
\item{table_name}{Character: Name of the supplementary table to remove}

\item{project_dir}{Character: Path to project directory (default: ".")}
}
\value{
Invisible TRUE on success
}
\description{
Removes a supplementary table and its metadata from a project
}
\examples{
\dontrun{
# Remove a supplementary table
removeRatTable("old_data", project_dir = "my_analysis")
}

}
